refactor(storage): extract row dataclasses to rows.py
PR-1 of the 3-PR repository.py split. Moves 12 row dataclasses and
WatchlistCardRow to storage/rows.py. repository.py re-exports them so
existing callers ('from uw_scan.storage.repository import JobRow') work
unchanged.

repository.py: 5173 -> ~5040 lines.
